requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
pywebview==4.1
pyinstaller==5.13.0